+++ plugins/record-hook/src/index.ts
@@ -1,15 +1,4 @@
-import { FlatfileListener, FlatfileEvent } from "@flatfile/listener";
-import { RecordHook } from "./RecordHook";
-import type { FlatfileRecord } from "@flatfile/hooks";
 export * from "./RecordHook";
+export * from "./record.hook.plugin";
 
-export const recordHook = (
-  sheetSlug: string,
-  callback: (record: FlatfileRecord) => {}
-) => {
-  return (client: FlatfileListener) => {
-    client.on("commit:created", { sheetSlug }, (event: FlatfileEvent) => {
-      return RecordHook(event, callback);
-    });
-  };
-};
+export { FlatfileRecord } from "@flatfile/hooks";