+++ plugins/xlsx-extractor/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./excel.extractor";
+export * from "./xlsx.extractor.plugin";